+++ src/lib/output/plugins/index.ts
@@ -2,3 +2,4 @@ export { MarkedPlugin } from "../themes/MarkedPlugin";
 export { AssetsPlugin } from "./AssetsPlugin";
 export { JavascriptIndexPlugin } from "./JavascriptIndexPlugin";
 export { NavigationPlugin } from "./NavigationPlugin";
+export { SitemapPlugin } from "./SitemapPlugin";