+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+## Features
+
+-   Added a new `--sitemapBaseUrl` option. When specified, TypeDoc will generate a `sitemap.xml` in your output folder that describes the site, #2480.
+
 ## v0.25.7 (2024-01-08)
 
 ### Bug Fixes